requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
av==11.0.0
azure-cognitiveservices-speech==1.36.0
Babel==2.14.0
bangla==0.0.2
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.3.1
confection==0.1.4
contourpy==1.2.0
coqpit==0.0.17
ctranslate2==4.1.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.9
dateparser==1.1.8
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
einops==0.7.0
elevenlabs==0.2.27
emoji==2.8.0
encodec==0.1.1
enum34==1.1.10
executing==2.0.1
faster-whisper==1.0.1
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
g2pkk==0.1.2
grpcio==1.62.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
h11==0.14.0
halo==0.0.31
hangul-romanize==0.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
inflect==7.0.0
ipython==8.22.2
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
log-symbols==0.0.14
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.26.4
onnxruntime==1.17.1
openai==1.13.3
openai-whisper==20231117
packaging==24.0
pandas==1.5.3
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
pure-eval==0.2.2
pvporcupine==1.9.5
PyAudio==0.2.14
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynndescent==0.5.12
pyparsing==3.1.2
pypinyin==0.51.0
pypiwin32==223
pyreadline3==3.4.1
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
RealTimeSTT==0.1.12
# Editable Git install with no remote (RealTimeTTS==0.3.43)
-e d:\projekte\realtimetts
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spinners==0.0.24
srsly==2.4.8
stable-ts==2.15.10
stack-data==0.6.3
stanza==1.6.1
stream2sentence==0.2.3
SudachiDict-core==20240109
SudachiPy==0.6.8
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
tqdm==4.66.2
trainer==0.0.36
traitlets==5.14.2
transformers==4.39.2
TTS==0.22.0
typer==0.9.4
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.5
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webrtcvad==2.0.10
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
